Update updated-questions.md
Heya - this is looking really great to me! I've proposed some changes to the formatting and wording so the layout is more readable. I think you may also need to add a few more words to the glossary potentially - but i like that it's clear instructions and everything in one place.